refactor(usb): move descriptor out from usb, mark usb gated by cargo feature
