+++ rmk/src/descriptor.rs
