+++ rmk/src/via/mod.rs
@@ -13,7 +13,7 @@ use crate::config::VialConfig;
 use crate::hid::{HidError, HidReaderTrait, HidWriterTrait};
 use crate::keymap::KeyMap;
 use crate::state::ConnectionState;
-use crate::usb::descriptor::ViaReport;
+use crate::descriptor::ViaReport;
 use crate::via::keycode_convert::{from_via_keycode, to_via_keycode};
 use crate::{boot, CONNECTION_STATE, MACRO_SPACE_SIZE};
 #[cfg(feature = "storage")]
